requirements.txt
fastapi
asyncpg
aiohttp
python-dotenv
